GEODE-3265: Refactoring MiscellaneousCommands
